Only update the record when fields have changed
For the token and metadata program add additional check to only update
the record if the values of the fields have changed and the slot is
larger so the slot column only changes when fields are updated. There
are accounts that are referencd in instructions which trigger account
update but they aren't changed. Every time update statement occurs
indexs are rewritten causing locks. This noop allows insert spam but
skips updating the record and thus no triggering of index updates.